Fixes #208 Histogram type is part of the public API (#210)
